pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "videoforge"
version = "0.1.0"
description = "Natural language video creation system powered by FFmpeg and AI"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [{ name = "VideoForge" }]

dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "click>=8.0",
    "pillow>=10.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
tts = ["voicevox-client>=0.4"]
ai = ["openai>=1.0", "stability-sdk>=0.8"]
all = ["videoforge[tts,ai]"]
dev = ["pytest>=8.0", "ruff>=0.5"]

[project.scripts]
videoforge = "videoforge.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/videoforge"]

[tool.ruff]
target-version = "py311"
line-length = 100
